requirements.txt
fastapi==0.136.0
uvicorn==0.44.0
sqlalchemy==2.0.49
pydantic==2.13.2
pydantic-core==2.46.2
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.2.2
python-multipart==0.0.26
email-validator==2.3.0
starlette==1.0.0
anyio==4.13.0
